moved OVR feeders
