+++ Customisations/OVRHandMeshDataFeeder.cs
@@ -1,11 +1,10 @@
 ﻿using System;
-using System.Reflection;
 using Assets.RemoteHandsTracking.Data;
 using Assets.RemoteHandsTracking.Extensions;
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace Assets.RemoteHandsTracking.Customisations.OVR
+namespace Assets.RemoteHandsTracking.Customisations
 {
     public class OVRHandMeshDataFeeder: MonoBehaviour
     {
